requirements.txt
playsound==1.3.0
paho-mqtt==1.6.1